Create symlinks under $GOFISH_BINPATH
Before this change, `gofish init` had been correctly using `$GOFISH_BINPATH` to override the default bin path(e.g. `/usr/local/bin` on *nix) to be created, but it was not using it when installing foods. This patch updates `gofish install` to create symlinks under `$GOFISH_BINPATH` if set, so that the behavior is consistent.